requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain_huggingface
sentence-transformers
tqdm
python-dotenv
huggingface-hub
transformers
faiss-cpu
requests
